pyproject.toml
[project]
name = "product-classification-hiring-demo"
version = "0.1.0"
description = "商品分类-测试（马上赢）"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "modelscope[framework]>=1.30.0",
    "jupyterlab>=4.4.9",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
    "tabulate>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
start = "src.main:main"
notebook = "jupyterlab.labapp:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -s"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
